Created index specific for last run
Before this all of the runs have been aggregated together. The
problem with this is that it can be difficult to only see results
for the latest run.  This change will create a new index for
each run and delete the old one.  So that the latest run
will be the only run available under this index.

Change-Id: Ic533bb5d4c0a80eb1dcbd462fb547889823fc292